Initial CI using GitHub Actions
Add an initial CI for testing the RPM build & installation of this repo
using GitHub Actions. The build job will setup the test systems, install
tito, and use it to test the RPM build & the installation of the
results (which are saved as artifacts).

Testing is done on the last Fedora release, and on CentOS Stream 8 & 9;
tito is available only in EPEL, so make sure to set it up on CentOS.

In addition, install tito manually using pip on CentOS Stream 9 from
the last version: this is because at the moment not all the dependencies
of tito are packaged [1], so tito is not installable.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2061761