Add build/packaging system
Add a simple Makefile that only "installs" the certificates to the
wanted location; it is all that is needed here, so a Makefile will
suffice for now.

Add a minimal RPM spec file, with bits mostly copied from
subscription-manager; the differences between the
subscription-manager-rhsm-certificates RPM currently generated by
subscription-manager and this spec file are:
- subscription-manager-rhsm-certificates is a noarch package now
- python-rhsm-certificates is no more provided, as the migration from it
  was done more than 4 years ago
- the license file is shipped as well, following more recent packaging
  practices

Add a configuration for tito to tag using the version pattern YYYYMMDD
rather than the classic X.Y.Z: this is done as X.Y.Z is basically
pointless for this repository (the files change once every few years),
so a date-based approach will work just fine.