CI: run build job automatically once a month
This repository does not see changes that often, and so if anything
breaks (usually in CI) we do not get to notice that soon enough.

Hence, run the build CI jobs automatically at 8:00 of the first day of
each month: it should be enough to notice whether there are new issues
showing up.